requirements.txt
asgiref==3.9.2
Django==5.2.6
sqlparse==0.5.3
geopy==2.4.1
Pillow==10.0.0
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1
drf-spectacular==0.27.0
